fix: inline launch logic, remove broken launch_common import
ROS 2 launch files in the install share directory aren't on sys.path,
so importing launch_common.py as a module fails. Inline the small
launch function in each launch file instead.
